Support Ranges
Originally, Ranges were not supported at all. This change implements a Redimension of the plotted trace if slicing is done from the second dimension with a fixed p value.

This needs some adjustments if the first dimension is the one defining the slice range. I keep that private until a final solution is done.